Port ApplicationEventHandler to Composer & Component
