+++ src/Articulate/Articulate.csproj
@@ -45,6 +45,8 @@
     <Compile Include="ArticulateSearchRouteHandler.cs" />
     <Compile Include="BlogMlExporter.cs" />
     <Compile Include="BlogMlImporter.cs" />
+    <Compile Include="Components\ArticulateComponent.cs" />
+    <Compile Include="Components\ArticulateComposer.cs" />
     <Compile Include="Controllers\ArticulateAuthorController.cs" />
     <Compile Include="Controllers\ArticulateArchiveController.cs" />
     <Compile Include="Controllers\ArticulateAuthorsController.cs" />
@@ -133,7 +135,6 @@
     <Compile Include="Syndication\RssFeedGenerator.cs" />
     <Compile Include="TagsOrCategoryPathRouteConstraint.cs" />
     <Compile Include="UmbracoHelperExtensions.cs" />
-    <Compile Include="UmbracoEventHandler.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ArticulateVirtualNodeByIdRouteHandler.cs" />
     <Compile Include="UrlHelperExtensions.cs" />
@@ -365,6 +366,7 @@
       <HintPath>..\packages\UmbracoCms.Web.8.0.0-alpha.58.1580\lib\net472\Umbraco.Web.UI.dll</HintPath>
     </Reference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\AutoMapper.3.3.1\tools\AutoMapper.targets" Condition="Exists('..\packages\AutoMapper.3.3.1\tools\AutoMapper.targets')" />
   <Import Project="..\packages\Umbraco.SqlServerCE.4.0.0.1\build\Umbraco.SqlServerCE.targets" Condition="Exists('..\packages\Umbraco.SqlServerCE.4.0.0.1\build\Umbraco.SqlServerCE.targets')" />
